v3 uerId身分驗證版本

+++ routes/api.php
@@ -4,7 +4,7 @@ use App\Http\Controllers\MainController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-const VERSION = 'v2';
+const VERSION = 'v3';
 
 /*
 |--------------------------------------------------------------------------
@@ -23,11 +23,17 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::prefix(VERSION)->group(function () {
     Route::prefix('/todos')->group(function () {
-        Route::get('/{id}', [MainController::class, 'getToDo']);
-        Route::get('', [MainController::class, 'getAllToDo']);
-        Route::post('', [MainController::class, 'newToDo']);
-        Route::put('/{id}', [MainController::class, 'updateToDo']);
-        Route::delete('/{id}', [MainController::class, 'deleteToDo']);
-        Route::delete('', [MainController::class, 'deleteAllToDo']);
+        Route::get('/{userId}/{id}', [MainController::class, 'getToDo']);
+        Route::get('/{userId}', [MainController::class, 'getAllToDo']);
+        Route::post('/{userId}', [MainController::class, 'newToDo']);
+        Route::put('/{userId}/{id}', [MainController::class, 'updateToDo']);
+        Route::delete('/{userId}/{id}', [MainController::class, 'deleteToDo']);
+        Route::delete('/{userId}', [MainController::class, 'deleteAllToDo']);
     });
+
+    // Route::prefix('/users')->group(function () {
+    //     Route::post('/register', [MainController::class, 'newToDo']);
+    //     Route::post('/login', [MainController::class, 'newToDo']);
+    //     Route::post('/logout', [MainController::class, 'newToDo']);
+    // });
 });
